Make Repository with Todo, FixedTodo, Calendar
